Change name of plugin, delete non used code, and fixed documentation and copyright.
