+++ site/plugins/archives_bar/archives_bar.plugin
@@ -1,0 +1,13 @@
+[Core]
+Name = archives_bar
+Module = archives_bar
+
+[Documentation]
+Author = Maria Andrea Vignau
+Version = 1.0
+Website = https://github.com/marian-vignau
+Description = archives_bar include renderer.
+
+[Nikola]
+PluginCategory = Task
+MinVersion = 7.8.2
