--- site/plugins/sidebar/templates/mako/sidebar-helper.tmpl
@@ -1,75 +1,0 @@
-## -*- coding: utf-8 -*-
-
-<%def name="list_posts(posts)">
-  <div id="global-recent-posts">
-    <h2>
-      ${ messages("Newer posts", lang) }
-    </h2>
-    <ul class="post-list">
-    % for post in posts:
-      <li>
-        <a href="${ post.permalink() }" class="listtitle">${ post.title(lang) }</a>
-      </li>
-    % endfor
-    </ul>
-  </div>
-</%def>
-
-<%def name="list_months(month_list)">
-  <div id="archive_month_list">
-    <select id="archive">
-      <option value="">Elija un mes ..</option>
-      % for month in month_list:
-      <option
-        value="${'/{}/{}'.format(month.year,month.month)}" class="listtitle">
-        ${ month.friendly_name }
-      </option>
-      % endfor
-    </select>
-   </div>
-</%def>
-
-<%def name="time_ago(year_ago_post)">
-  <div id="year_ago_post">
-      <a href="link://${ year_ago_post.permalink() }" class="listtitle">${ year_ago_post.title(lang) }</a>
-   </div>
-</%def>
-
-<%def name="list_taxonomy(title, items=None, hierarchy=None, type='tag')">
-  % if items or hierarchy:
-        <div id="global-${ type }-list">
-          <h2>
-            ${ messages(title, lang) }
-          </h2>
-    % if hierarchy:
-      % for text, full_name, path, link, indent_levels, indent_change_before, indent_change_after, children_count, count in hierarchy:
-        % for i in range(indent_change_before):
-          <ul class="${ type }-level ${ type }-level-${ len(indent_levels) + i }">
-        % endfor
-            <li><a class="reference" href="${ link }">${ text|h }</a> (${ count })
-        % if indent_change_after <= 0:
-            </li>
-        % endif
-        % for i in range(-indent_change_after):
-          </ul>
-          % if i + 1 < len(indent_levels):
-          </li>
-          % endif
-        % endfor
-      % endfor
-    % elif items:
-          <ul>
-        % for text, count, link in items:
-          % if text:
-            <li><a class="reference" href="${ link }">${ text|h }</a> (${ count })</li>
-          % endif
-        % endfor
-          </ul>
-    % endif
-        </div>
-  % endif
-</%def>
-
-<%def name="archives(archives)">
-${ list_taxonomy('Archive', hierarchy=archives, type='archive') }
-</%def>
