--- site/plugins/sidebar/templates/mako/sidebar.tmpl
@@ -1,37 +1,0 @@
-## -*- coding: utf-8 -*-
-<script src="https://code.jquery.com/jquery-1.12.4.js"></script>
-<script>
-    document.getElementById("archive").onchange = function() {
-        if (this.selectedIndex!==0) {
-            window.location.href = this.value;
-        }
-    };
-</script>
-
-<%namespace name="helper" file="sidebar-helper.tmpl"/>
-<div>
-  <h2>
-    Un año atrás
-  </h2>
-  ${ helper.time_ago(year_ago_post) }
-</div>
-
-<h2>
-Archivo
-</h2>
-${ helper.list_months(month_list) }
-
-<!--
-% if global_posts:
-  ${ helper.list_posts(global_posts) }
-% endif
-% if global_category_items:
-  ${ helper.list_taxonomy("Categories", global_category_items, global_category_hierarchy, type='category') }}
-% endif
-% if global_archive_hierarchy:
-  ${ helper.archives(global_archive_hierarchy) }
-% endif
-% if global_tag_items:
-  ${ helper.list_taxonomy("Tags", global_tag_items, type='tag') }
-% endif
--->
