+++ site/themes/blogtxt/templates/base.tmpl
@@ -54,8 +54,8 @@ ${template_hooks['extra_head']()}
             <div id="propaganda_c">
             </div><!-- to include a random propaganda -->
 
-            <div id="sidebar_inner">
-            </div><!-- to include sidebar generated by plugin -->
+            <div id="archive_bar">
+            </div><!-- to include archive month list generated by plugin -->
             <div id="tag_cloud">
             </div><!-- to include tag cloud generated by plugin -->
             <ul>
@@ -100,7 +100,7 @@ ${template_hooks['extra_head']()}
         ${base.late_load_js()}
     <script>
     $( "#propaganda_c" ).load( "${rel_link(permalink,'/propaganda_c.html')}" );
-    $( "#sidebar_inner" ).load( "${rel_link(permalink,'/sidebar-es.html')}" );
+    $( "#archive_bar" ).load( "${rel_link(permalink,'/archive_bar-es.html')}" );
     $( "#tag_cloud" ).load( "${rel_link(permalink,'/tagcloud-es.inc.html')}" );
     </script>
     <script src="${rel_link(permalink,'/assets/js/printer-friendly.js')}"></script>
